% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidPlyr-package}
\alias{AcidPlyr}
\alias{AcidPlyr-package}
\title{AcidPlyr}
\description{
A dplyr-like grammar for manipulation of S4 rectangular data classes, such
as \code{DFrame}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/acidplyr/}
  \item \url{https://github.com/acidgenomics/r-acidplyr/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-acidplyr/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Authors:
\itemize{
  \item Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})
}

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
