% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cast-methods.R
\name{cast}
\alias{cast}
\alias{cast,DataFrame-method}
\title{Cast data from long to wide format}
\usage{
cast(object, ...)

\S4method{cast}{DataFrame}(object, colnames = "colname", values = "value")
}
\arguments{
\item{object}{Object.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Pivot column data from long format to wide format.
}
\note{
Updated 2022-09-16.
}
\examples{
data(matrix, package = "AcidTest")

## DataFrame ====
df <- melt(matrix)
}
\seealso{
\itemize{
\item https://seananderson.ca/2013/10/19/reshape/
\item \code{reshape2::dcast} (deprecated).
\item \code{tidyr::pivot_wider}.
\item Python \code{pandas.pivot_table}.
\item Python \code{pandas.wide_to_long}.
\item Python \code{pandas.DataFrame.pivot}.
}
}
