% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/collapseToString-methods.R
\name{collapseToString}
\alias{collapseToString}
\alias{collapseToString,atomic-method}
\alias{collapseToString,matrix-method}
\alias{collapseToString,data.frame-method}
\alias{collapseToString,DataFrame-method}
\title{Collapse to string}
\usage{
collapseToString(object, ...)

\S4method{collapseToString}{atomic}(object, sep = ", ", sort = FALSE, unique = FALSE)

\S4method{collapseToString}{matrix}(object, sep = ", ", sort = FALSE, unique = FALSE)

\S4method{collapseToString}{data.frame}(object, sep = ", ", sort = FALSE, unique = FALSE)

\S4method{collapseToString}{DataFrame}(object, sep = ", ", sort = FALSE, unique = FALSE)
}
\arguments{
\item{object}{Object.}

\item{sep}{\code{character(1)}.
Separator. Defaults to comma.}

\item{sort}{\code{logical(1)}.
Sort values.}

\item{unique}{\code{logical(1)}.
Unique values.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{atomic}: \code{character(1)}.
\item \code{dim}: Object of same class, collapsed to a single row.
}
}
\description{
Collapse to string
}
\note{
Updated 2021-09-03.
}
\examples{
## character ====
groceries <- c(NA, NA, "milk", "eggs", "eggs", "veggies")
collapseToString(
    groceries,
    unique = TRUE,
    sort = TRUE
)
collapseToString(
    groceries,
    unique = FALSE,
    sort = FALSE
)

## numeric ====
collapseToString(seq(1:5))

## logical ====
collapseToString(c(TRUE, FALSE))
collapseToString(c(NA, NaN))

## data.frame ====
df <- datasets::iris
df <- head(df)
df <- collapseToString(df, sort = TRUE, unique = TRUE)
t(df)
}
\seealso{
\code{toString()}.
}
