% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/filterNested-methods.R
\name{filterNested}
\alias{filterNested}
\alias{filterNested,DFrame,character-method}
\title{Keep rows that match a nested condition}
\usage{
filterNested(object, pattern, ignoreCase, ...)

\S4method{filterNested}{DFrame,character}(object, pattern, ignoreCase = FALSE)
}
\arguments{
\item{object}{Object.}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}

\item{ignoreCase}{\code{logical(1)}.
Perform case insensitive pattern matching.}

\item{...}{Additional arguments.}
}
\value{
Modified object, containing only matching rows.
}
\description{
Keep rows that match a nested condition
}
\details{
Intended for partial matching with regular expressions. For exact matches,
refer to \code{matchNested} function.
}
\note{
Updated 2023-12-12.
}
\examples{
## DFrame ====
pattern <- "^[A-Z]+$"
object <- S4Vectors::DataFrame(
    "V1" = c("a", "b", "c"),
    "V2" = I(list(
        c("aa", "bb", "CC"),
        c("dd", "ee", "ff"),
        c("gg", "hh", "ii")
    ))
)
object <- filterNested(object = object, pattern = pattern)
print(object)
}
