% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/join-methods.R
\name{innerJoin}
\alias{innerJoin}
\alias{leftJoin}
\alias{rightJoin}
\alias{fullJoin}
\alias{semiJoin}
\alias{antiJoin}
\alias{join}
\alias{antiJoin,DFrame,DFrame,character-method}
\alias{fullJoin,DFrame,DFrame,character-method}
\alias{innerJoin,DFrame,DFrame,character-method}
\alias{leftJoin,DFrame,DFrame,character-method}
\alias{rightJoin,DFrame,DFrame,character-method}
\alias{semiJoin,DFrame,DFrame,character-method}
\title{Join operations}
\usage{
innerJoin(x, y, by, ...)

leftJoin(x, y, by, ...)

rightJoin(x, y, by, ...)

fullJoin(x, y, by, ...)

semiJoin(x, y, by, ...)

antiJoin(x, y, by, ...)

\S4method{antiJoin}{DFrame,DFrame,character}(x, y, by)

\S4method{fullJoin}{DFrame,DFrame,character}(x, y, by)

\S4method{innerJoin}{DFrame,DFrame,character}(x, y, by)

\S4method{leftJoin}{DFrame,DFrame,character}(x, y, by)

\S4method{rightJoin}{DFrame,DFrame,character}(x, y, by)

\S4method{semiJoin}{DFrame,DFrame,character}(x, y, by)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{by}{\code{character}.
Column names to use for join operation.}

\item{...}{Additional arguments.}
}
\value{
New object containing a merge of \code{x} and \code{y} objects.
}
\description{
Join operations
}
\note{
Updated 2021-10-14.
}
\section{Ordering}{


The join functions never rearrange rows. To accomplish this, we're currently
setting an internal \code{.idx} column that we can use to reorder the rows after
\code{merge()} operation.
}

\section{Row names}{


Unlike the S3 methods defined in dplyr, the join methods defined here for
\code{DFrame} always preserve row names.
}

\examples{
data(band_members, band_instruments, package = "AcidTest")

## DFrame ====
x <- as(band_members, "DFrame")
print(x)
y <- as(band_instruments, "DFrame")
print(y)
by <- "name"
innerJoin(x = x, y = y, by = by)
leftJoin(x = x, y = y, by = by)
rightJoin(x = x, y = y, by = by)
fullJoin(x = x, y = y, by = by)
semiJoin(x = x, y = y, by = by)
antiJoin(x = x, y = y, by = by)
}
\seealso{
These functions are inspired by dplyr. However, they are designed to only
work on Bioconductor S4 class objects, and use base R code internally.
\itemize{
\item \code{help(topic = "join", package = "dplyr")}.
\item \code{help(topic = "merge", package = "base")}.
\item https://support.bioconductor.org/p/120277/
}
}
