% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/melt-methods.R
\name{melt}
\alias{melt}
\alias{melt,DataFrame-method}
\alias{melt,matrix-method}
\alias{melt,table-method}
\title{Melt data from wide to long format}
\usage{
melt(object, ...)

\S4method{melt}{DataFrame}(object, colnames = c("rowname", "colname", "value"))

\S4method{melt}{matrix}(
  object,
  colnames = c("rowname", "colname", "value"),
  min = -Inf,
  minMethod = c("absolute", "perRow"),
  trans = c("identity", "log2", "log10")
)

\S4method{melt}{table}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{colnames}{\code{character(3)}.
Column name mappings for melted data frame return.}

\item{min}{\code{numeric(1)} or \code{NULL}.
Minimum count threshold to apply. Filters using "greater than or equal to"
logic internally. Note that this threshold gets applied prior to
logarithmic transformation, when \code{trans} argument applies.
Use \code{-Inf} or \code{NULL} to disable.}

\item{minMethod}{\code{character(1)}.
Only applies when \code{min} argument is numeric.
Uses \code{match.arg()}.
\itemize{
\item \code{absolute}: Applies hard cutoff to \code{counts} column after the melt
operation. This applies to all counts, not per feature.
\item \code{perRow}: Applies cutoff per row (i.e. gene). Internally, \code{rowSums()}
values are checked against this cutoff threshold prior to the melt
operation.
}}

\item{trans}{\code{character(1)}.
Apply a log transformation (e.g. \code{log2(x + 1L)}) to the count matrix prior
to melting, if desired. Use \code{"identity"} to return unmodified (default).}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Unpivot column data from wide format to long format.
}
\note{
Updated 2021-09-03.
}
\examples{
data(matrix, package = "AcidTest")

## matrix ====
dim(matrix)
x <- melt(matrix)
dim(x)
print(x)
}
\seealso{
\itemize{
\item https://seananderson.ca/2013/10/19/reshape/
\item \code{reshape2::melt} (deprecated).
\item \code{tidyr::pivot_longer}.
\item Python \code{pandas.melt}.
}
}
