% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutate-methods.R
\name{mutateAll}
\alias{mutateAll}
\alias{mutateAt}
\alias{mutateIf}
\alias{transmuteAt}
\alias{transmuteIf}
\alias{mutate}
\alias{mutateAll,DFrame,function-method}
\alias{mutateAt,DFrame,character,function-method}
\alias{mutateIf,DFrame,function,function-method}
\alias{transmuteAt,DFrame,character,function-method}
\alias{transmuteIf,DFrame,function,function-method}
\title{Mutate multiple columns}
\usage{
mutateAll(object, fun, ...)

mutateAll(object, fun, ...)

mutateIf(object, predicate, fun, ...)

transmuteAt(object, vars, fun, ...)

transmuteIf(object, predicate, fun, ...)

\S4method{mutateAll}{DFrame,function}(object, fun, ...)

\S4method{mutateAt}{DFrame,character,function}(object, vars, fun, ...)

\S4method{mutateIf}{DFrame,function,function}(object, predicate, fun, ...)

\S4method{transmuteAt}{DFrame,character,function}(object, vars, fun, ...)

\S4method{transmuteIf}{DFrame,function,function}(object, predicate, fun, ...)
}
\arguments{
\item{object}{Object.}

\item{fun}{\code{function}.
Mutation function.}

\item{...}{Passthrough arguments to function declared in \code{fun} argument.}

\item{vars}{\code{character}.
Column names.}

\item{predicate}{\code{function}.
A predicate function to be applied to the columns.
For example, \code{is.atomic()}.}
}
\value{
Modified object.
}
\description{
Mutate multiple columns
}
\note{
Updated 2023-02-07.
}
\examples{
data(mtcars, package = "datasets")

## DFrame ====
x <- as(mtcars, "DFrame")
mutateAll(x, fun = log, base = 2L)
mutateAt(x, vars = c("mpg", "cyl"), fun = log, base = 2L)
mutateIf(x, predicate = is.double, fun = as.integer)
transmuteAt(x, vars = c("mpg", "cyl"), fun = log, base = 2L)
transmuteIf(x, predicate = is.double, fun = as.integer)
}
\seealso{
These functions are inspired by dplyr. However, they are designed to only
work on Bioconductor S4 class objects, and use base R code internally.
}
