% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest-methods.R
\name{nest}
\alias{nest}
\title{Nest rows of a single column into a list-column}
\arguments{
\item{col}{\code{character(1)}.
Name or position of the column to nest into a list.}

\item{...}{Additional arguments.}
}
\description{
Nest rows of a single column into a list-column
}
\note{
Updated 2023-08-23.
}
\examples{
## DFrame ====
long <- melt(matrix)
print(long)
wide <- cast(long)
print(wide)
}
