% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/rbindToDataFrame-methods.R
\name{rbindToDataFrame}
\alias{rbindToDataFrame}
\alias{rbindToDataFrame,list-method}
\title{Row bind elements 1:1 to a data frame}
\usage{
rbindToDataFrame(x, ...)

\S4method{rbindToDataFrame}{list}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Row bind elements 1:1 to a data frame
}
\note{
Updated 2023-02-22.
}
\examples{
x <- list(
    "a" = list(
        "aa" = seq(from = 1L, to = 3L),
        "bb" = seq(from = 4L, to = 6L)
    ),
    "b" = list(
        "cc" = seq(from = 7L, to = 9L),
        "dd" = seq(from = 10L, to = 12L)
    ),
    "c" = list(
        "ee" = seq(from = 13L, to = 15L),
        "ff" = seq(from = 16L, to = 18L)
    )
)
print(x)
y <- rbindToDataFrame(x)
print(y)
}
\seealso{
\itemize{
\item \code{data.table::rbindlist}.
\item \code{purrr::list_transpose}.
}
}
