% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{DataFrame}
\alias{map_df}
\alias{map_dfr}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidGenerics}{\code{\link[AcidGenerics:reexports]{DataFrame}}}

  \item{purrr}{\code{\link[purrr:map]{map_df}}, \code{\link[purrr:map]{map_dfr}}}
}}

