% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/unlist2-methods.R
\name{unlist2}
\alias{unlist2}
\alias{unlist2,DFrameList-method}
\title{Flatten lists}
\usage{
unlist2(x, ...)

\S4method{unlist2}{DFrameList}(x, nameCol = "name", rownameCol = "rowname")
}
\arguments{
\item{x}{Object.}

\item{nameCol}{\code{character(1)}.
Column name to assign \code{names} of input object.}

\item{rownameCol}{\code{character(1)}.
Column name to assign \code{rownames} of input object.}

\item{...}{Additional arguments.}
}
\value{
Unlisted object derived from original input.
For example, \code{DFrame} derived from \code{DFrameList}.
}
\description{
Flatten lists
}
\details{
Intended primarily for complex S4 classes, such as \code{DFrameList}.
Intentionally does not support recursion.
}
\note{
Updated 2023-11-30.
}
\examples{
## DFrameList ====
suppressPackageStartupMessages({
    library(IRanges)
})
object <- as.DataFrame(list(
    "col1" = CharacterList(
        c("a", "b", "c", "d"),
        c("e", "f", "g"),
        c("h", "i")
    ),
    "col2" = IntegerList(
        seq(from = 1L, to = 2L),
        seq(from = 3L, to = 4L),
        seq(from = 5L, to = 6L)
    ),
    "col3" = c("a", "b", "c")
))
print(object)
x <- unnest2(object, col = "col1")
print(x)
y <- unnest2(object, col = "col2")
print(y)
}
\seealso{
\itemize{
\item \code{unlist()}.
}
}
