% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/unlistToDataFrame-methods.R
\name{unlistToDataFrame}
\alias{unlistToDataFrame}
\alias{unlistToDataFrame,list-method}
\title{Unlist to a data frame}
\usage{
unlistToDataFrame(x, ...)

\S4method{unlistToDataFrame}{list}(x, recursive = TRUE)
}
\arguments{
\item{x}{Object.}

\item{recursive}{\code{logical(1)}.
\itemize{
\item \code{TRUE}: Recursively unlist all nested list columns.
Calls \code{purrr::map_dfr} internally.
\item \code{FALSE}: Only unlists the top level of list, allowing for retention
of nested list columns and/or complex S4 objects.
}}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Unlist to a data frame
}
\note{
Updated 2021-02-19.
}
\examples{
x <- list(
    a = list(
        aa = seq(from = 1L, to = 3L),
        bb = seq(from = 4L, to = 6L)
    ),
    b = list(
        cc = seq(from = 7L, to = 9L),
        dd = seq(from = 10L, to = 12L)
    ),
    c = list(
        ee = seq(from = 13L, to = 15L),
        ff = seq(from = 16L, to = 18L)
    )
)
print(x)
y <- unlistToDataFrame(x, recursive = TRUE)
y <- unlistToDataFrame(x, recursive = FALSE)
print(x)
}
\seealso{
\itemize{
\item \code{purrr::map_dfr}.
\item \code{plyr::ldply} (deprecated).
}
}
