% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/unnest2-methods.R
\name{unnest2}
\alias{unnest2}
\alias{unnest2,DFrame-method}
\title{Unnest a single list-column into expanded rows}
\usage{
unnest2(object, ...)

\S4method{unnest2}{DFrame}(object, col)
}
\arguments{
\item{col}{\code{character(1)}.
Name of the list-column to unnest into long format.}

\item{...}{Additional arguments.}
}
\description{
Unnest a single list-column into expanded rows
}
\note{
Updated 2023-08-23.
}
\examples{
## DFrame ====
}
