% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/unlist2-methods.R
\name{unlist2}
\alias{unlist2}
\alias{unlist2,DFrameList-method}
\title{Flatten lists}
\usage{
unlist2(x, ...)

\S4method{unlist2}{DFrameList}(x, nameCol = "name", rownameCol = "rowname")
}
\arguments{
\item{x}{Object.}

\item{nameCol}{\code{character(1)} or \code{NULL}.
Column name to assign \code{names} of input object.}

\item{rownameCol}{\code{character(1)} or \code{NULL}.
Column name to assign \code{rownames} of input object.}

\item{...}{Additional arguments.}
}
\value{
Unlisted object derived from original input.
For example, \code{DFrame} derived from \code{DFrameList}.
}
\description{
Flatten lists
}
\details{
Intended primarily for complex S4 classes, such as \code{DFrameList}.
Intentionally does not support recursion.
}
\note{
Updated 2024-01-04.
}
\examples{
## DFrameList ====
suppressPackageStartupMessages({
    library(IRanges)
})
dfl <- DataFrameList(
    "A" = DataFrame(
        "a" = c(1L, 2L),
        "b" = c(3L, 4L),
        row.names = c("aa", "bb")
    ),
    "B" = DataFrame(
        "a" = 5L,
        "b" = 6L,
        row.names = "cc"
    )
)
## Our variant, which assigns into columns.
x <- unlist2(dfl)
print(x)
## Compare with the standard variant, which modifies rownames.
y <- unlist(dfl)
print(y)
}
\seealso{
\itemize{
\item \code{unlist()}.
}
}
