% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/unlistToDataFrame-methods.R
\name{unlistToDataFrame}
\alias{unlistToDataFrame}
\alias{unlistToDataFrame,list-method}
\title{Unlist to a data frame}
\usage{
unlistToDataFrame(x, ...)

\S4method{unlistToDataFrame}{list}(x, mode = c("map_dfr", "base"))
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Unlist to a data frame
}
\note{
Updated 2021-02-19.
}
\examples{
x <- list(
    "a" = list(
        "aa" = seq(from = 1L, to = 3L),
        "bb" = seq(from = 4L, to = 6L)
    ),
    "b" = list(
        "cc" = seq(from = 7L, to = 9L),
        "dd" = seq(from = 10L, to = 12L)
    ),
    "c" = list(
        "ee" = seq(from = 13L, to = 15L),
        "ff" = seq(from = 16L, to = 18L)
    )
)
print(x)
y <- unlistToDataFrame(x)
print(y)
}
\seealso{
\itemize{
\item \code{purrr::map_dfr()}.
\item \code{plyr::ldply()} (\emph{deprecated}).
\item \code{pipette::as.DataFrame()}.
}
}
