% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/unnest2-methods.R
\name{unnest2}
\alias{unnest2}
\alias{unnest2,DFrame-method}
\title{Unnest a single list-column into expanded rows}
\usage{
unnest2(object, ...)

\S4method{unnest2}{DFrame}(object, col)
}
\arguments{
\item{object}{Object.}

\item{col}{\code{character(1)}.
Name of the list-column to unnest into long format.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with a single unnested list-column expanded into
rows.
}
\description{
Unnest a single list-column into expanded rows
}
\note{
Updated 2023-08-24.
}
\examples{
## DFrame ====
suppressPackageStartupMessages({
    library(IRanges)
    library(pipette)
})
object <- as.DataFrame(list(
    "col1" = CharacterList(
        c("a", "b", "c", "d"),
        c("e", "f", "g"),
        c("h", "i")
    ),
    "col2" = IntegerList(
        seq(from = 1L, to = 2L),
        seq(from = 3L, to = 4L),
        seq(from = 5L, to = 6L)
    ),
    "col3" = c("a", "b", "c")
))
print(object)
x <- unnest2(object, col = "col1")
print(x)
y <- unnest2(object, col = "col2")
print(y)
}
\seealso{
\itemize{
\item \code{tidyr::unnest()}.
}
}
